% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onewayAMaxT.R
\name{onewayAMaxT}
\alias{onewayAMaxT}
\title{Find test statistic value from a given data and corresponding critical value}
\usage{
onewayAMaxT(data, k, alpha)
}
\arguments{
\item{data}{a real matrix}

\item{k}{positive integer}

\item{alpha}{real number between 0 and 1 called significance level}
}
\value{
numeric vector
}
\description{
More detailed description
}
\examples{
k=3;N=c(20,10,10);S=c(1,2,6)
g=NULL
for(i in 1:k)
{
 g[[i]]=rnorm(N[i],0,sqrt(S[i]))
}
data=g
onewayAMaxT(data,k,0.05)
}
