% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrderedTests.R
\docType{package}
\name{OrderedTests}
\alias{OrderedTests}
\title{Find test statistic value from a given data and corresponding critical value}
\description{
It gives test statistic and critical value for testing homogeneity of treatment effects against ordered alternatives in one-way ANOVA, two-way ANOVA and one-way ANCOVA model.
}
\author{
Anjana Mondal \email{mondal.anjana25@gmail.com}
}
